feat: add populate method to getAllProducts
This commit populates reviews into products using virtuals to reference
each other. While it is possible to add a direct reference to reviews
into the product model, it is considered bad practice.A product can have multiple reviews, so you would need to create an array field for reviews in the Product model. This can become difficult to manage as the number of reviews grows, and probably would also become really inefficient for querying or updating.

Separating all reviews into their own Review model is the better approach. This allows for better management and querying of reviews, and allows you to add more information to each review if needed. By separating reviews into their own model, you can also create a more complex schema for reviews if necessary, with fields such as timestamps, likes, or categories. Using virtuals is a good method for relating to models when needed.